Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[argon2]
google-generativeai
argon2-cffi